+++ CHANGELOG.md
@@ -5,10 +5,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.5.6] - 2018-10-07
 ### Added
+- Reset option @AndrewRedican.
 - Russian translation for warnings @askesis.
 - French translation for warnings @cbonaudo.
 - Bahasa Indonesia translation for warnings @radityasurya.
+
+### Fixed
+- Issue where component did not compare new and previous `placeholder` object correctly. 
 
 ## [2.5.5] - 2018-09-19
 ### Added